Fixed version code

+++ BmwDeepObd/ActivityCommon.cs
@@ -4429,7 +4429,7 @@ namespace BmwDeepObd
                         formDownload.Add(new StringContent(string.Format(CultureInfo.InvariantCulture, "{0}",
                             packageInfo != null ? Android.Support.V4.Content.PM.PackageInfoCompat.GetLongVersionCode(packageInfo) : 0)), "appver");
                         formDownload.Add(new StringContent(GetCurrentLanguage()), "lang");
-                        formDownload.Add(new StringContent(string.Format(CultureInfo.InvariantCulture, "{0}", Build.VERSION.SdkInt )), "android_ver");
+                        formDownload.Add(new StringContent(string.Format(CultureInfo.InvariantCulture, "{0}", (long) Build.VERSION.SdkInt )), "android_ver");
                         formDownload.Add(new StringContent(Build.Fingerprint), "fingerprint");
                         formDownload.Add(new StringContent(SelectedInterface.ToDescriptionString()), "interface_type");
                         formDownload.Add(new StringContent(LastAdapterSerial ?? string.Empty), "adapter_serial");
@@ -4514,7 +4514,7 @@ namespace BmwDeepObd
                     sb.Append("Deep OBD Trace info");
                     sb.Append(string.Format("\nDate: {0:u}", DateTime.Now));
                     sb.Append(string.Format("\nLanguage: {0}", GetCurrentLanguage()));
-                    sb.Append(string.Format("\nAndroid version: {0}", Build.VERSION.SdkInt));
+                    sb.Append(string.Format("\nAndroid version: {0}", (long) Build.VERSION.SdkInt));
                     sb.Append(string.Format("\nAndroid id: {0}", Build.Id ?? string.Empty));
                     sb.Append(string.Format("\nAndroid display: {0}", Build.Display ?? string.Empty));
                     sb.Append(string.Format("\nAndroid fingerprint: {0}", Build.Fingerprint ?? string.Empty));
@@ -5229,7 +5229,7 @@ namespace BmwDeepObd
                         packageInfo != null ? Android.Support.V4.Content.PM.PackageInfoCompat.GetLongVersionCode(packageInfo) : 0)), "app_ver" },
                     { new StringContent(AppId), "app_id" },
                     { new StringContent(GetCurrentLanguage()), "lang" },
-                    { new StringContent(string.Format(CultureInfo.InvariantCulture, "{0}", Build.VERSION.SdkInt)), "android_ver" },
+                    { new StringContent(string.Format(CultureInfo.InvariantCulture, "{0}", (long) Build.VERSION.SdkInt)), "android_ver" },
                     { new StringContent(Build.Fingerprint), "fingerprint" },
                     { new StringContent(installer ?? string.Empty), "installer" },
                     { new StringContent(SelectedInterface.ToDescriptionString()), "interface_type" },
